# services/internal/user/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Documentation from User Service",
        "title": "User App Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/user",
    "paths": {
        "/login": {
            "post": {
                "description": "test deneme",
                "produces": [
                    "application/json"
                ],
                "summary": "Login With Given Credentials.",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginUserRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回用户信息",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginUserResponseDTO"
                        }
                    },
                    "400": {
                        "description": "err_code：10002 参数错误； err_code：10003 校验错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "err_code：10001 登录失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "err_code：20001 服务错误；err_code：20002 接口错误；err_code：20003 无数据错误；err_code：20004 数据库异常；err_code：20005 缓存异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "@Me Endpoint.",
                "produces": [
                    "application/json"
                ],
                "summary": "Gives All Tokens with Refresh.",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回用户信息",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerSuccessResponse-dto_WhoAmIResponseDto-any"
                        }
                    },
                    "400": {
                        "description": "err_code：10002 参数错误； err_code：10003 校验错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "err_code：10001 登录失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "err_code：20001 服务错误；err_code：20002 接口错误；err_code：20003 无数据错误；err_code：20004 数据库异常；err_code：20005 缓存异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register Endpoint.",
                "produces": [
                    "application/json"
                ],
                "summary": "Register With Given Credentials.",
                "parameters": [
                    {
                        "description": "Enter Credential.",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterUserRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok\" \"返回用户信息",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerSuccessResponse-any-any"
                        }
                    },
                    "400": {
                        "description": "err_code：10002 参数错误； err_code：10003 校验错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "err_code：10001 登录失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "err_code：20001 服务错误；err_code：20002 接口错误；err_code：20003 无数据错误；err_code：20004 数据库异常；err_code：20005 缓存异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.LoginUserRequestDto": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.LoginUserResponseDTO": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "dto.RegisterUserRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "lastname",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.SwaggerPayload-any-any": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {}
            }
        },
        "dto.SwaggerPayload-dto_WhoAmIResponseDto-any": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/dto.WhoAmIResponseDto"
                }
            }
        },
        "dto.SwaggerSuccessResponse-any-any": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/dto.SwaggerPayload-any-any"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.SwaggerSuccessResponse-dto_WhoAmIResponseDto-any": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/dto.SwaggerPayload-dto_WhoAmIResponseDto-any"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.WhoAmIResponseDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "X-Token",
            "in": "header"
        }
    }
}